added BoundingBox to Directions

+++ src/main/java/at/fhtw/app/viewModel/TourFormViewModel.java
@@ -34,7 +34,7 @@ public class TourFormViewModel {
                 formTour.getToLocation(),
                 formTour.getTransportType(),
                 directions.getDistance(),
-                (double) directions.getTime(), null);
+                directions.getTime(), null);
         //send Tour entries to server
         System.out.println("Form Tour: " + formTour.toString());
